feature: Integracion del sistema de ecepciones

+++ src/main/java/com/vhre/auth_boilerplate/modules/roles/service/RoleService.java
@@ -1,5 +1,6 @@
 package com.vhre.auth_boilerplate.modules.roles.service;
 
+import com.vhre.auth_boilerplate.core.exception.ResourceNotFoundException;
 import com.vhre.auth_boilerplate.modules.roles.dto.RoleRequestDTO;
 import com.vhre.auth_boilerplate.modules.roles.dto.RoleResponseDTO;
 import com.vhre.auth_boilerplate.modules.roles.entity.Role;
@@ -19,9 +20,13 @@ public class RoleService {
     private final RoleRepository roleRepository;
     private final RoleMapper roleMapper;
 
+    @Transactional
     public RoleResponseDTO create(RoleRequestDTO dto) {
-        if (roleRepository.existsByName(dto.name())) throw new RuntimeException("Role exists");
-        return roleMapper.toResponse(roleRepository.save(roleMapper.toEntity(dto)));
+        if (roleRepository.existsByName(dto.name())) {
+            throw new IllegalArgumentException("Role already exists with name: " + dto.name());
+        }
+        Role role = roleMapper.toEntity(dto);
+        return roleMapper.toResponse(roleRepository.save(role));
     }
 
     public List<RoleResponseDTO> findAll() {
@@ -29,17 +34,24 @@ public class RoleService {
     }
 
     public RoleResponseDTO findById(UUID id) {
-        return roleMapper.toResponse(roleRepository.findById(id).orElseThrow());
+        return roleMapper.toResponse(roleRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Role not found with ID: " + id)));
     }
 
     @Transactional
     public RoleResponseDTO update(UUID id, RoleRequestDTO dto) {
-        Role role = roleRepository.findById(id).orElseThrow();
+        Role role = roleRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Role not found with ID: " + id));
+
         role.setName(dto.name());
         return roleMapper.toResponse(roleRepository.save(role));
     }
 
+    @Transactional
     public void delete(UUID id) {
+        if (!roleRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Role not found with ID: " + id);
+        }
         roleRepository.deleteById(id);
     }
 }
